Don't exceed MAX_JOBS_RUNNING when we get extra matches. #4554
When we send a resource request list to the negotiator, we may get back
enough matches to exceed MAX_JOBS_RUNNING. When this happens, the
ScheddNegotiate object will skip job ids in the autocluster and then
call scheduler_handleMatch() with job ids of -1.-1, which then finds
another job for the match and starts it.
The end result is that the schedd starts too many shadows, then kills
some to come back under MAX_JOBS_RUNNING.

Ensure that scheduler_handleMatch() will call skipAllSuchJobs() with a
valid job id in this case to catch that MAX_JOBS_RUNNING has been
reached.