Implement ENCRYPT_EXECUTE_DIRETORY on Linux. #4856
Encrypt the execute directory on linux via ecryptfs and
private mount namespaces. If enabled, the startd will create
a subdirectory in the execute dir, and while spawning the starter,
the startd will perform an ecryptfs mount of this subdirectory
in a mount namespace private to the starter.  The starter will
then create the scratch directory for this job on this ecryptfs mount.