Allow jobs to request encrypt_execute_directory in submit file. #4856
In addition honoring admin desires to encrypt the execute directory
via a knob in condor_config, allow users to specify it in the
submit file on both Windows and Linux.  This also requires the starter
to advertise it as a capability via machine attr HasEncryptExecuteDirectory,
and for condor_submit to adjust the machine requirements accordingly.