Add special attribute CurrentTime to ClassAds library. #4858
When old ClassAds semantics are enabled, CurrentTime becomes a special
attribute that evaluates to the current unix epoch time, and is
equivalent to calling the ClassAd function time().

Adding this special handling will mean that Condor code will no longer
need to add CurrentTime=time() to every ad it uses in order to support
the old ClassAds behavior, which is still relied upon in many places.