Aggressively rewrite shared port addresses. #4848 #4849
If SHARED_PORT_ADDRESS_REWRITING is set to true, THEN:

- Check address rewriting for all command sockets.  If the ad contained
an IPv6 (perhaps because the collector connected to itself over
the default IP) but the default command socket is IPv4, then address
rewriting did not work.  We now iterate through all known command
sockets.  In addition, when the shared port is used, we must
communicate the list of possible command sockets through the
shared_port_ad.

- If we see an sinful on the same shared port server, but not the
same shared port ID as our command socket, we still rewrite the socket.