Have the schedd generate a restart reconnect report. #4783
The schedd can now write a report to disk summarizing the reconnect on
restart statistics. The report file is updated every minute until all
reconnect attempts are finished. At that point, the final report is
emailed to the pool admin.

The new config param SCHEDD_RESTART_REPORT controls where the report is
written. If the value is empty, then no report is written or emailed.