Avoid matchmaking with schedds near limits. #4905
Previously, the scheduler would send out very large resource requests
(in the wild, we observed 70,000 jobs in an autocluster) despite being
at-or-near the MAX_JOBS_RUNNING limit.  This tracks how close we are
to the limit and trims resource requests accordingly.

Additionally, this introduces the concept of CURB_MATCHMAKING, a config
expression which allows us to prevent matchmaking in admin-defined
situations.