Make tear-down methods safe to call after DaemonCore is deleted. #4909
We may have global objects that try to call DaemonCore cancel methods in
their destructors. There's one in the starter that cancels a timer for
user job policy evaluation. But the daemonCore object will already be
gone, so the process can crash.
To minimize this possiblity, turn the DaemonCore cancel methods into
noops if the 'this' pointer is NULL.