Write job events with a single write(). #4908
We want to disable locking of the job event logs by default. To do this,
we need to ensure that each complete event is written with a single
write(), and the file is opened in append mode (except when updating the
header event of the event log on rotation).
We buffer the contents of the log event ourselves, instead of relying on
libc to do so.