Don't use temporary local file for the event log rotation lock. #4908
Use a permanent, long-lived lock file for the event log rotation lock.
This means that the schedd, shadow, and gridmanager don't need to
acquire a write lock and delete the lock file every time they delete a
WriteUserLog object.
The default lock file is $(LOCK)/EventLogLock, which should be on local
disk.