Read consecutive messages from messenger.  #4928
If multiple CEDAR messages are available on a socket, it will take
a DC duty cycle per message.  This may be inefficient if each message
takes a short time to process.

With this change, multiple queued messages will be processed at once
(as long as the total runtime is less than 2 seconds).