Restore gnu linker option to call real exit() in our wrapper. #4910
On unix, we replace the libc exit() function because we don't want to
perform normal at-exit cleanup when a daemon forks off a child helper
process (without exec()). In the move to cmake, we lost the option to
call the real exit() when possible (using a gnu linker option).
This change restores it.