revert 0130317e105 on Win32, but leave it in place for other platforms
When DefaultMakeClassAdLogTableEntry is declared in the classad_log.h header,
it leads to a crash on at process termination on some linux platforms.  This is
fixed by moving the declaration to classad_log.cpp.  But the fix on linux
introduces a crash a process termination on Windows.  So the previous code
is back inside a #ifdef WIN32 block.