Fix time zone handling in python bindings for ClassAd AbsTime. #4936
When converting a Python datetime value to a ClassAd AbsTime value, the
bindings should calculate the local time zone offset from UTC based on
the time of the value, not the current time. Otherwise, a timestamp
value won't round-trip properly if the value is during standard time and
the current time is DST, or vice versa.

We can use the same function that the ClassAds libraries uses to make
the calculation.

Conflicts:
	src/python-bindings/classad.cpp