Discard all buffered data when ReliSock is closed or deleted. #4915
Add reset() method to SndMsg and RcvMsg to discard all buffers,
including the SndMsg:m_out_buf buffer that holds data during
non-blocking write operations (and also fix bug where this buffer
is potentially leaked when a ReliSock is deallocated). Invoke reset()
whenever a sock is closed, so if the same sock is used for another
connection, we don't pollute the new connection with old buffered data!
Implemented by making close() virtual and implementing close() inside
ReliSock and SafeSock that discards the buffers and then invokes
parent class close to do the rest.