Ignore MATCH_INFO for pslots with consumption policies. #4945
With consumption policies, it doesn't make sense for the pslot to be in
Matched state, since it can be matched to many jobs at the same time.
And due to how claims for consumption policies were implemented,
reception of MATCH_INFO messages makes some of the claims unusable and
can leave the slot in Matched state forever.