Eliminate dead code
This is cleanup work for #4958

This function was used by runOldLogic, which was just debugging and
development code that was no longer needed and tended to spam the logs
with useless warnings.  runOldLogic is gone, so this function has no
users, so it's time go to, not least of which because an unused function
is a warning and warnings are errors on several BATLab platforms.