Clarify debug message when a function does not exist. #4885
When a function does't exist, the debug() invocation simply
reports ERROR.  This is difficult to differentiate between
"function evaluated to the constant ERROR" versus "function does
not exist".  It now logs 'ERROR (function does not exist)'.

Internally, I had to add a new boolean method to the FunctionCall
class; I did not change the vtable, so no new revision to the ABI.