Revert "Revert "First batch of submit foreach tests which test what results from #4954""
This reverts commit 6f9b3912c6c9153583ac60fe820accaa76065b84.

The biggest change is the multiMonitor needed for foreach submits
with multiple logs. We check using condor_submit -dry-run
when we parse the submit file. This data is stored and we look at both
the number of logs and the number of jobs. We don't really care how
jobs go witth each userlog. However more then one log has us call
the multiMonitor which forks a monitor for each log and then waits for
them all.

We had a serious timeout on windows and reverted this. The problem
was correctly extracting the std error -dry-run data which will be fixed
before pushing this revert revert.