Fix wait-forever bug when socket closed while waiting to read/write #4997
When the Selector class is used to wait for a single socket to become
available for reading or writing, we use poll() but fail to check for
the POLLHUP result, which indicates that the connection has been closed.
This can result in the caller waiting forever.
We've observed this happening in condor_ssh_to_job.