Add modified slot back to match list when using CP. #4973
Previously, when a slot was sent to a schedd, it was removed
from the MatchList -- this means a p-slot will be ignored until
the match list is recomputed (typically, when a new auto-cluster
is considered).

Now, we subtract out the resources that will be used by the d-slot
and add the p-slot back to the match list.