Re-sort match list after re-adding p-slot. #4973
When using consumption policies, we re-add the smaller p-slot
back into the match list.  With this commit, we re-sort the
match list (using insertion sort).

This places the new p-slot in the appropriate match list location
with respect to ranking policies.  For example, if we have:

NEGOTIATOR_PRE_JOB_RANK = Cpus

then the smaller p-slot may have a lower rank (and hence be lower
in the match list) after it is matched once.

This patch was tested by setting up a personal HTCondor instance with
4 partitionable slots, each containing 4 cores, and NEGOTIATOR_PRE_JOB_RANK
as above.  A cluster of 8 jobs was submitted; with this patch, each
p-slot received two matches (instead of 2 p-slots receiving 4 matches).