Revamp ClassAd parsing. #4990
Replace old, inconsistent parsing methods (parse, parseOld, parseOldAds)
with a set of functions that provide better semantics.

- parseOne: Returns a single ad from input, even if the input has multiple
  ads.
- parseNext: Consume the next classad in the input, returning the ad and
  advancing the input file-like object.
- parseAds: Returns an iterator that will parse the ClassAds.

The above take a keyword argument to specify which parser to use (old/new),
but default to auto-detecting the format.

Consistent with other parts of HTCondor, garbage lines are ignored by the
old ClassAd parser (this helps in the parsing of output from 'condor_q -g -l').

Use of the old functions trigger a python deprecation warning; this can
be disabled by setting ENABLE_DEPRECATION_WARNINGS=false in the HTCondor
configuration.

See ticket for details.