Starter can wait forever when connection to shadow closes. #5136
When the handler indicating that the connection to the shadow has
probably closed is invoked, use the whoami RPC instead of calling
updateShadow(). The latter won't attempt to use the connection if the
job has completed, resulting in the starter not starting its timer to
abort after the job lease time elapses.

Also return KEEP_STREAM, so that daemon-core doesn't delete the
ReliSock. Subsequent starter code may reference it in its closed state.