Unconditionally remove SUBMIT_Iwd from routed job. #5181
The job router users the presence of SUBMIT_Iwd to detect sandboxed
jobs when reconnecting to jobs post-restart.

However, if the job was not sandboxed in the first place - and it
was remotely submitted - it will have SUBMIT_Iwd present in the
source and routed jobs.  Hence, on restart, JobRouter assumes that
all routed jobs are sandboxed.  This results in the JR trying to
do file transfer on all reconnected jobs - which will fail because
the SUBMIT_Iwd points at a directory on the remote submitter's host.

This removes SUBMIT_Iwd for all routed jobs, sandboxed-or-not.  Since
in the non-sandboxed case we won't attempt a file xfer, it shouldn't
hurt.