When shadow is exiting quickly, tell the starter to do the same. #5121
When the shadow is doing a quick "hold job and exit" operation, it
should always request a fast shutdown of the starter, instead of the
default graceful shutdown request. With a graceful shutdown, the starter
may wait around for a dead shadow to reconnect.
This is usually caused by a bad file access or an unrecoverable error
(like file transfer failure) reported by the starter.