Differentiate location lookups from queries. #5150
The collector fundamentally has two roles: responding to queries
for pool information (i.e., condor_status) and providing a DNS-like
service for daemons (as is used in condor_q -pool pool.example.com -name foo).

Previously, there was no differentiation between the two types of queries -
the response and work for the collector was the same in both cases.  Now,
location queries apply a strict projection on the attributes returned.

Further, the collector is notified that this is a "location query" - in the
future it may be able to further optimize the query server-side (for example,
by maintaining a special index for location lookups).