Reduce chowning of job spool directories. #5226
Add a new (default) option to have job files in the SPOOL directory
spend most/all of their time owned by the job owner. This eliminates
the chowning of files that currently happens before the job starts
running and after it completes.
When the new config param CHOWN_JOB_SPOOL_FILES is False (the default),
job files are created as the job owner and maintain that ownership for
their entire lifetime.
Chowning may still occur when upgrading from an older version or when
the value of CHOWN_JOB_SPOOL_FILES is changed.

The SOAP interface needs more work. Input files are created with owner
condor and then chowned when the submission is committed, as the schedd
may not know the job's owner when the files are created. Files are read
as the condor user, which may fail if files have retrictive read
permissions.