Support relative path install; non-list JSON
Our JSON file is not properly formed; it's a bunch of items, when what we
need is a single item (that being a list of the other items).  render.js
now handles that.

The reliance on foo.cgi was removed to simplify things.

Paths were adjusted to support installs nearly anywhere, including file://
URLs.