The schedd shouldn't set match diagnostic attributes for bogus job id. #5270
If the schedd asks for matches for jobs that subsequently can't use them
(e.g. MAX_JOBS_RUNNING would be exceeded), the negotiation code will use
a job id of -1.-1 as the candidate job for the extra matches.
If the negotiator sends a REJECTED message, then the schedd could set
match rejection attributes for this non-existent job.
This didn't bother the schedd, but could confuse anyone else reading its
job queue log. The job router in particular was tripped up by this.