Fix blahp problem when job IDs are reused by the batch system. #5062
Some batch systems will reuse job IDs. If an ID gets reused when the
blahp's job registry still has an entry for the previous use of that ID,
the status of the two jobs gets conflated. If old registry entries are
purged too quickly, then the blahp can lose track of jobs that are still
in the queue.

This patch applies a simple fix, where jobs with a status that indicates
they are still in the queue are not purged at the normal age limit.
Thus, the purge lifetime can be adjusted to be significantly shorter
than the expected ID reuse interval without losing track of long-running
jobs.