don't use automatic dependencies for external-libs #5349
The automatically generated Provides from condor-external-libs overlap
with the real globus packages (etc), and the automatic Requires for
the main condor package, which for UW_BUILD are intended to refer to
condor-external-libs, then get fulfilled by the system-installed ones.

This disables automatic Provides generation for condor-external-libs
and uses a custom find-requires script to filter out those libs where
they are required from the main condor package (and the cream-gahp
subpackage).  Instead, we explicitly require the external-libs package.
%{?_isa} is included to make sure the right arch for external-libs is
pulled in.