Add a resource request prefetch phase to the negotiator. #5174
If enabled in the config file, the negotiator will concurrently connect
to each schedd that has one or more active submitters and start send one
NEGOTIATE per submitter ad to each schedd, serially. After reading the
requests for one submitter and not returning any matches, it will start
a new NEGOTIATE for the next submitter. After the prefetch phase, when
it gets to each submitter in the normal blocking cycle, it will send
another NEGOTIATE command, followed by a stream of matches for the
prefetched requests. If it runs out of prefetched requests, it will send
SEND_JOB_INFO as needed as before.