Condor-C jobs aborted when lease is less than one hour #5446
The batched lease renewal code in the gridmanager can now be told to
reevaluate the next time a lease needs to be renewed (e.g. when a new
job is submitted). Without this, a new job with a short lease duration
may not get its lease renewed when necessary. In particular, a lease
duration under one hour may not be renewed before it expires.