Constrain and validate accounting group names. #5538
If an accounting group name contains a space, the negotiator will write
invalid entries in the accountant log and be unable to read the log back
in on startup.
The negotiator now ignores submitter ads with an invalid name and
condor_submit fails if given an invalid accounting group or user name.