Make Target scoping option in MatchClassAd with old semantics. #5550
When old ClassAd evaluation semantics are enabled, MatchClassAd will now
set alternateScope in the left and right ads to point at each other.
The compatibility layer in Condor did this before, but the Job Router
bypasses the compatibility layer in several places, leading to
inconsistent rules on when Target scoping is required.