GRIDMANAGER_SELECTION_EXPR causes overwriting grid ads #5683
When GRIDMANAGER_SELECTION_EXPR is used, the selection value needs to be
included in the gridmanager ads and used as part of the hash key in the
collector. Otherwise, ads from different gridmanagers can overwrite
each other.
Added new attribute GridmanagerSelectionValue and new command line
argument to the gridmanager to communicate this value.