Pslot preemption can preempt wrong dslots. #5748
Don't set PreemptDslotClaims in the machine ad until we want to send
the ad to a schedd. If we set it while building the MatchList and then
don't use the ad for the current submitter, we could send the ad to a
subsequent submitter with an inappropriate set of dslot claims when that
match didn't require preemption.