Use proc_pid_rusage() on mac for procapi, where available. #5777
task_for_pid() is unreliable for obtaining memory and cpu usage of
processes, more so in Mac OS 10.11 and above. SIP means it fails for
processes based on system binaries, while spewing errors to the system
logs.
Mac OS 10.9 introduced proc_pid_rusage(), which gives us the information
we need. It's unaffected by SIP and non-root processes can use it on
processes with the same owner.