On linux, remove destination file when installing shared libraries.
This affects developers running HTCondor out of a build tree.
We unconditionally copy and strip shared libraries from the externals on
a 'make install'. On linux, if a running daemon is using the libraries
under release_dir and a new 'make install' is done, then the shared
library files mapped into the process have their contents altered,
resulting in a crash the next time they are accessed.