Improve recovery from procd failure in daemons. #5863
When the procd dies and restarts, other daemons need to reopen their
pipes to it. This happens automatically in ProcFamilyProxy for most
requests to the procd. But it can't happen for requests that occur
during Create_Process(). This is because recovery may involve calling
Create_Process() to spawn the procd, and the forked child process may be
sharing the parent's memory, complicating recovery in the child.
As a result, all calls to Create_Process() will fail until the daemon
does a different procd request.

If a procd failure is detected during Create_Process(), we now have the
parent process register a one-shot timer in DaemonCore to do a simple
procd request. This will trigger the recovery code in ProcFamilyProxy.
The Create_Process() call will still fail, but subsequent ones will not
suffer the same problem.