Work around bogus read() results when reading /proc/pid/environ #5863
We are seeing instances where read() returns values in the 1GB when
asked to read the contents of /proc/<pid>/environ into a 1MB buffer.
Attempts to re-read the file (with and without reopening it) give the
same result.
In this situation, give up without trying to extract environment
information about the process. Before, we would attempt to examine
large amounts of random data, resulting in a crash.