Allow user to supply additional parameters for GCE instance request. #5893
New submit parameter gce_json_file (job ad attribute GceJsonFile) allows
the user to provide a file containing JSON text that should be included
in the instance description that the gce_gahp sends to the GCE server.
No attempt is made to merge elements that overlap or conflict with
elements that the gce_gahp defines in the instance description.