Fix credential filenames when using accounting groups. #5964
Credential files in SEC_CREDENTIAL_DIRECTORY are named after the job
owner. The schedd was using submitter names to detect when a credential
should be removed. When accounting groups are being used, the submitter
name is different than the owner of a job.
Now, the schedd uses recently-introduced per-owner data structures to
detect when an owner has no jobs queued and to name the .mark file
propriately.