Add singularity bind mount support.  #6011
This introduces a new parameter, SINGULARITY_BIND_EXPR, into the
configuration file for the starter.  When set, it is evaluated as
a ClassAd expression with the TARGET set to the job ad and MY set
to the machine ad.

The resulting string is parsed as a list of bind mount
specifications that are subsequently passed to Singularity.

For example, to force jobs to always mount /cvmfs inside the
container, one may set:

SINGULARITY_BIND_EXPR = "/cvmfs"