fd_set bounds checking can cause abort in Selector class #6061
On some configurations of linux, FD_SET() will abort if given a file
descriptor value that will index past the end of one fd_set structure.
We do this safely in the Selector class by using any array of fd_sets.
To avoid triggering the bounds check, we now wrap FD_SET() (and friends)
with a macro that calculates the address of the desired fd_set within
the array of fd_sets and the offset within that fd_set.