On Windows, build condor_utils library faster with multiple cores.  #6079
We do this by setting the /MP flag when building source files in
condor_utils, so we get task-level parallelism when the utils lib
is built on its own first.  We do not want to use
the /MP other modules, since we already are using project-level parallelism
once the utils are built.