modify test files to include the test requirements
Currently the test requirements are mentioned in the
'Test_Requirement' file. For example the first entry
in the file is:

cmd_q_shows-better-analyze: personal

This implies that the test cmd_q_shows-better-analyze
requires a personal condor to be running, similarly
there are entries for tests which have other requirements.
These entries are read by by run_test.pl/batch_test.pl
scripts which are used to run the unit tests and if the
test requirements are not satisfied they throw an error.

This change depracates 'Test_requirements' file. The
requirements will now be present in the test file itself
as a comment in the following format:

^#testreq: <space separated requirements>
(^ indicates that the comment should start from beginning
of the line. ^ is not present in the actual file)

ex. for cmd_q_shows-better-analyze.run
^#testreq: personal

Furthermore, if a test requires a personal condor to be
running, run_test.pl will start condor just like the nmi
scripts for batlab, instead of throwing an error.

Few other tests, start condor themseleves using boilerplate
code. To address this problem, we have added support for
tests to specify test configuration in a here document marked
by CONDOR_TESTREQ_CONFIG. Such a here document is parsed only
if it is between the ^#testreq and ^#endtestreq markers.

^#testreq: personal
my $config = <<CONDOR_TESTREQ_CONFIG;
    DAEMON_LIST = MASTER,SCHEDD,COLLECTOR,NEGOTIATOR,STARTD
    WANT_SUSPEND = FALSE
    NEGOTIATION_INTERVAL = 5
    SCHEDD_INTERVAL = 5
    NUM_CPUS = 24
    MAX_SCHEDD_LOG = 20 Kb
CONDOR_TESTREQ_CONFIG
^#endtestreq

ex. check the new job_test_scheddrotation.run

run_test.pl and nmi scripts have been modified to parse the
requirements and configuration from test files themselves,
batch_test.pl still uses the depracated method and should not be used.