Fix Job Router's handling of job event logs. #6092
Events for the routed job are no longer written to the event logs of the
source job, nor is the source job's event logs overwritten when the
routed job completes. The Job Router now correctly handles relative
pathnames when writing events. The global event log is now correctly
written to when the job doesn't request any event logs.