for #6130, speed up ingesting classads into the collector by
Added a new getClassAdEx function to condor_utils, this function has options that allow for
   * optimized parsing and insert of literal attributes
   * lazy parsing via the classadCache
   * flags to control clearing the ad before parsing, and whether or not mytype and targettype are expected.

Added new optimized methods for inserting literal values into a classad, which getClassAdEx uses.
Added methods to cedar that return both a pointer and a length for strings and for secrets, so we can avoid copying data out of cedar buffers before parsing.

The only current user of getClassAdEx is the collector, when it handles an update command, in the future, we expect more users.

This commit also makes the unparsed form in the classad cache the same form that we put on the wire. "old" classad form