Enable proxy refresh between gridmanager and ft-gahp. #5841
Forwarding of a refreshed proxy can not be done between the
gridmanager and ft-gahp. It work similarly to file transfer between the
two daemons. The new ft-gahp command, DOWNLOAD_PROXY, takes a ClassAd
that contains a proxy file, server contact information and a capability.
The ft-gahp contacts the given sever and initiates the new
FETCH_PROXY_DELEGATION daemoncore command. After it presents the
capability, a proxy delegation occurs.
The gridmanager accepts the new FETCH_PROXY_DELEGATION command.
The result of the DOWNLOAD_PROXY command includes the path to the job
sandbox directory, into which the proxy was written.
The gridmanager needs this in order to tell the blahp about the
refreshed proxy.