Change default locate() method from LOCATE_FULL to LOCATE_FOR_LOOKUP. #6223
To locate call sites that needed changing and to prevent future
land-mines, we remove the public method Daemon::daemonAd() which is
used to retrieve the full classad after LOCATE_FULL. Call sites that
want to invoke the daemonAd() method must now instantiate a child
class DaemonAllowLocateFull instead of using class Daemon.
DaemonAllowLocateFull operates just like Daemon, except it includes a
daemonAd() method and the locate() method must be specified. This way,
the code is future-proofed in that if someone edits the code to
invoke method daemonAd() on a Daemon object, they will end up with a
compile-time error.