Update x509UserProxyExpiration attribute in Job Router. #6209
Propagate any changes to x509UserProxyExpiration from the source job ad
in the schedd to the ad used for job policy evaluation (in the RoutedJob
object). This attribute can change during the lifetime of the job and is
sometimes used in job policy (to remove jobs with an expired proxy).