Fix bug where condor_history ad may contain RemoteHost. #6251
Job attributes like RemoteHost are normally removed whenever the shadow associated with the job exits.
However, these attributes must be preserved to enable a new shadow to reconnect to the job in the
event the schedd is restarted. However, in the situation where the shadow exits because the schedd is
doing a graceful shutshutting down, previous to this patch the schedd
would  always preserve these attributes even if the job completed!
The result was the reconnect attributes
appeared in the history file whenever a job completed while the schedd
was in the process of a graceful shutdown.