Change negotiator naming to match naming for master and schedd. #5717
The default negotiator name has been changed back to the hostname. Added
command line argument -name to set the name. If a name is set via -name,
-local-name or NEGOTIATOR_NAME, the value is passed through
build_valid_daemon_name() to determine the actual name.
This is consistent with the naming of master and schedd daemons, and
works better with Condor's assumption that a daemon name without a '@'
is a fully-qualified hostname.