Split gen_ckpt_name() calls between two new functions. #6221
Building the path to a spooled executable, or ickpt, file is now done
with GetSpooledExecutablePath(), which requires only the cluster id.

Building the path to a job's spool directory (or checkpoint file, for
standard universe) is done with SpooledJobFiles::getJobSpoolPath(),
which takes a job ad. Its result may be affected by the new alternate
spool directory feature.
The old form of getJobSpoolPath() is kept for soap, where the job ad
isn't available when creating the job's spool directory.