Add ALTERNATE_JOB_SPOOL config knob #6221
ALTERNATE_JOB_SPOOL is an expression evaluated against the job ad. If it
evaluates to a string, that string is used as an alternate value for
SPOOL when determining the job's spool directory or checkpoint file.
The alternate directory is assumed to exist and have the same
ownership/permissions as the main SPOOL directory.
This expression is evaluated each time
SpooledJobFiles::getJobSpoolPath() is called, and there's no graceful
handling if the value for a given job changes over the job's lifetime.