Enforce use of the OS version of python on our mac build machines. #6244
Some of our MacOSX build machines have an alternate version of python
installed in /usr/local/bin, which appears early in the default PATH.
This version may be older than the python library in /usr/lib, causing
our cmake configuration to disable the python bindings.
For now, add /usr/bin to the front of the PATH on these machines, so
that the system python is always used.