Change how we set dynamic library paths for MacOSX executables. #6244
When creating a shared library, use the full path to the library for its
-install_name, rather than its basename.
Call install_name_tool to rename paths of dependent libraries at install
time, rather than at build time. The relative paths we set don't make
sense for for the binary files sitting in the build tree.

These two changes allow us to run the binaries that are sitting in the
build tree, before they're assembled into the install directory. This is
useful for the unit tests.
This can also be done by setting DYLD_LIBRARY_PATH, but that doesn't
work when testing the python bindings. The system's python interpreter
is protected by SIP, so DYLD_LIBRARY_PATH is cleared.