Strip path from mac shared libraries' id in the install step #6244
We recently changed our shared libraries so that their -install_name
contains their full path within the build space. That simplifies
building and running unit tests that use the libraries.
But once we stage the libraries outside of the build space, we should
strip the path information out of the -install_name. It's useless and
potential confusing everywhere else the libraries may end up going.