Restore the old dynamic library path rewriting for test programs on mac #6244
The test programs don't have an install directive. They're just written
into src/condor_tests. When a test run is done in the batlab, a symlink
is made so that ../lib points to the necessary lib directory. And since
the tests are run at a different time and in a different location than
the build, the full pathnames for the libraries won't be valid. So we
want to do the library path rewritting at compile time for these
programs, instead of at install time.