Fix negotiation protocol error when doing request prefetch. #6312
When prefetch sees NO_MORE_JOBS from the schedd, don't ask for more
request ads during normal negotiation. This was happening inadvertantly
when the last of the prefetched request ads were tossed because they
were in an autocluster that previously failed to be matched. As a
result, the negotiator would send an END_NEGOTIATE at the wrong time,
causing the schedd to close the conneciton.