Fix race condition between job removal and shadow recycling. #6353
Once the schedd decides to make a shadow exit (done via signal or CEDAR
message to the shadow), don't allow that shadow to recycle and fetch a
new job. There's a race condition where the signal can be applied to the
wrong job. The shadow can connect to the schedd to recycle after the
schedd starts sending the signal but before the shadow processes it. In
the worst case, this results in the newly fetched job being removed from
the queue (along with the old job).