(#6326) go out of our way to make FILE pos match the fd's
In Python 3, our FILE pointer does not correspond to anything in the
python object.  Buffered getc calls will advance the fd pos a block at a
time, and ungetc does not actually seek backwards.

Avoiding buffering keeps the FILE's pos in sync with the fd's, but this
is costly as each getc now corresponds to a separate read system call.

Is there a better way to handle this?