(#6326) re-open tempfile objects for python3
classad parsing does not seem to work on tempfile objects, but it does
for the files returned by open.

Apparently this is because tempfile.TemporaryFile() returns a
BufferedRandom object, which does not actually do a seek on the
underlying file descriptor when you call tf.seek(), but open() returns a
TextIOWrapper, which does in fact seek the fd when you call tf.seek().

This is relevant because the classad parser will start/continue from the
current position of the FILE object it's handed.