Revert "(#6326) drop unnecessary error state preservation"
This reverts commit 3a04619db16cdf5225628f0ec034e2eba96cd745.

Well, this is a bit confusing, but it appears that the PyErr_Fetch +
PyErr_Restore here are somehow relevant, even though followed by
PyErr_Clear.  At least on python2, this was causing parseNext to
prematurely return a StopIteration and the test_old_classad_v2 test in
classad_tests.py to fail.

Dunno what gives, but the problem goes away with this commit reverted.