Always try to send alive messages to systemd on linux. #6385
Previously, this code was only enabled if the relevant header and
library file were present at compile time. This caused a problem for a
user on ubuntu 14, where systemd isn't part fo the normal configuration,
but can be enabled. We don't need the development files at compile time,
so we should always enable this code on linux.