Make HAD replication work with shared port and IPv6. #6413
Add new file replication command where the files are sent over the
same connection. With the old command, the downloader bound a random
port on IPv4 and sent its address along with the command. The uploader
then connected to that dynamic port to do the data transfer. This
doesn't work for shared port or IPv6.

With the new command, the uploading transferer daemon inherits the
connection from its replication daemon parent, and no second
connection is made.

The replication daemon will fall back to using the old command if
talking to an old peer.