Allow CERequirements to be specified in a simpler way. #6133
The gridmanager now understands two simpler ways to specify name/value
pairs in the CERequirements job attribute used by the blahp. The
gridmanager will translate these forms into the goofy ClassAd expression
that the blahp expects.

If CERequirements evaluates to a nested ClassAd, then the attribute
names and evaluated values are used to build the expression for the
blahp.
If CERequirements evaluates to a string, the string is split on commas
and spaces to produce a list of attribute names, which are looked up in
the main job ad to obtain values to build the expression for the blahp.
Otherwise, CERequirements is forwarded unevaluated to the blahp.