Have systemd do a fast shutdown instead of sigkill when stopping. #6096
To get a fast shutdown when stopping the HTCondor service, we need to
tell systemd to send a SIGQUIT signal to just the condor_master process
when stopping the service. Note that we cannot use the systemd ExecStop
directive accomplish this, because condor_off is asynchronous and
systemd assumes synchronous -- as soon as the ExecStop command returns,
systemd immediately kills all daemons with SIGKILL. So instead,
we need to not specify ExecStop and instead use KillMode and
KillSignal to instruct systemd to send SIGQUIT to the
condor_master.